Part2/Part2_Backend/requirements.txt
Flask
python-dotenv
cloudinary
flask-cors
supabase
zappa
openpyxl
